reference_implementations/uncertainty-molecules/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
absl-py=0.14.1=pypi_0
aiohttp=3.7.4.post0=pypi_0
argon2-cffi=21.1.0=pypi_0
astunparse=1.6.3=pypi_0
async-timeout=3.0.1=pypi_0
attrs=21.2.0=pypi_0
backcall=0.2.0=pypi_0
blas=1.0=mkl
bleach=4.1.0=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.10.8=ha878542_0
cachetools=4.2.4=pypi_0
certifi=2021.10.8=py38h06a4308_0
cffi=1.15.0=pypi_0
chardet=4.0.0=pypi_0
charset-normalizer=2.0.7=pypi_0
click=8.0.3=pypi_0
colorama=0.4.4=pypi_0
configparser=5.0.2=pypi_0
cudatoolkit=11.1.74=h6bb024c_0
cycler=0.10.0=pypi_0
debugpy=1.5.0=pypi_0
decorator=5.1.0=pypi_0
defusedxml=0.7.1=pypi_0
deprecation=2.1.0=pypi_0
docker-pycreds=0.4.0=pypi_0
docopt=0.6.2=pypi_0
entrypoints=0.3=pypi_0
ffmpeg=4.4.0=hca11adc_0
flatbuffers=2.0=pypi_0
freetype=2.10.4=h5ab3b9f_0
fsspec=2021.10.0=pypi_0
future=0.18.2=pypi_0
gast=0.5.3=pypi_0
giflib=5.2.1=h7b6447c_0
gitdb=4.0.7=pypi_0
gitpython=3.1.24=pypi_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=2.3.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
googledrivedownloader=0.4=pypi_0
gpytorch=1.5.1=pypi_0
grpcio=1.41.0=pypi_0
h5py=3.6.0=pypi_0
idna=3.3=pypi_0
ignite=1.1.0=pypi_0
importlib-resources=5.4.0=pypi_0
intel-openmp=2021.3.0=h06a4308_3350
ipykernel=6.5.0=pypi_0
ipython=7.29.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.6.5=pypi_0
isodate=0.6.0=pypi_0
jedi=0.18.0=pypi_0
jinja2=3.0.2=pypi_0
joblib=1.1.0=pypi_0
jpeg=9b=h024ee3a_2
jsonpickle=1.5.2=pypi_0
jsonschema=4.2.1=pypi_0
jupyter-client=7.0.6=pypi_0
jupyter-core=4.9.1=pypi_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-widgets=1.0.2=pypi_0
keras=2.8.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.2=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libclang=13.0.0=pypi_0
libffi=3.4.2=h295c915_0
libgcc-ng=11.2.0=h1d223b6_11
libgomp=11.2.0=h1d223b6_11
libidn2=2.3.2=h7f8727e_0
libnsl=2.0.0=h7f98852_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=11.2.0=he4da1e4_11
libtasn1=4.17.0=h7f98852_0
libtiff=4.2.0=h3942068_0
libunistring=0.9.10=h27cfd23_0
libuv=1.42.0=h7f98852_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
libzlib=1.2.11=h36c2ea0_1013
lz4-c=1.9.3=h295c915_1
markdown=3.3.4=pypi_0
markupsafe=2.0.1=pypi_0
matplotlib=3.4.3=pypi_0
matplotlib-inline=0.1.3=pypi_0
mistune=0.8.4=pypi_0
mkl=2021.3.0=h06a4308_520
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.0=py38h42c9631_2
mkl_random=1.2.2=py38h51133e4_0
mpmath=1.2.1=pypi_0
multidict=5.2.0=pypi_0
munch=2.5.0=pypi_0
nbclient=0.5.8=pypi_0
nbconvert=6.3.0=pypi_0
nbformat=5.1.3=pypi_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.1=pypi_0
nettle=3.7.3=hbbd107a_1
networkx=2.6.3=pypi_0
ninja=1.10.2=hff7bd54_1
notebook=6.4.5=pypi_0
numpy=1.21.2=py38h20f2e39_0
numpy-base=1.21.2=py38h79a1101_0
oauthlib=3.1.1=pypi_0
olefile=0.46=pyhd3eb1b0_0
openh264=2.1.1=h780b84a_0
openssl=3.0.0=h7f98852_1
opt-einsum=3.3.0=pypi_0
packaging=21.0=pypi_0
pandas=1.3.3=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.2=pypi_0
pathtools=0.1.2=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=8.3.1=py38h5aabda8_0
pip=22.0.3=pypi_0
prometheus-client=0.12.0=pypi_0
promise=2.3=pypi_0
prompt-toolkit=3.0.21=pypi_0
protobuf=3.18.1=pypi_0
psutil=5.8.0=pypi_0
ptyprocess=0.7.0=pypi_0
py-cpuinfo=8.0.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pypi_0
pydeprecate=0.3.1=pypi_0
pygments=2.10.0=pypi_0
pymongo=3.12.0=pypi_0
pyparsing=2.4.7=pypi_0
pyqt5=5.15.6=pypi_0
pyqt5-qt5=5.15.2=pypi_0
pyqt5-sip=12.9.1=pypi_0
pyrsistent=0.18.0=pypi_0
python=3.8.12=hf930737_2_cpython
python-dateutil=2.8.2=pypi_0
python_abi=3.8=2_cp38
pytorch=1.9.1=py3.8_cuda11.1_cudnn8.0.5_0
pytorch-cluster=1.5.9=py38_torch_1.9.0_cu111
pytorch-ignite=0.4.7=pypi_0
pytorch-lightning=1.4.9=pypi_0
pytz=2021.3=pypi_0
pyyaml=6.0=pypi_0
pyzmq=22.3.0=pypi_0
rdflib=6.0.2=pypi_0
readline=8.1=h27cfd23_0
requests=2.26.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.7.2=pypi_0
sacred=0.8.2=pypi_0
scikit-learn=1.0=pypi_0
scipy=1.7.1=pypi_0
seml=0.3.4=pypi_0
send2trash=1.8.0=pypi_0
sentry-sdk=1.4.3=pypi_0
setuptools=58.0.4=py38h06a4308_0
shortuuid=1.0.1=pypi_0
six=1.16.0=pyhd3eb1b0_0
smmap=4.0.0=pypi_0
sqlite=3.36.0=hc218d9a_0
subprocess32=3.5.4=pypi_0
sympy=1.9=pypi_0
tensorboard=2.8.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
tensorflow=2.8.0=pypi_0
tensorflow-io-gcs-filesystem=0.24.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.12.1=pypi_0
testpath=0.5.0=pypi_0
tf-estimator-nightly=2.8.0.dev2021122109=pypi_0
threadpoolctl=3.0.0=pypi_0
tk=8.6.11=h1ccaba5_0
torch-geometric=2.0.1=pypi_0
torch-scatter=2.0.8=pypi_0
torch-sparse=0.6.12=pypi_0
torchaudio=0.9.1=py38
torchmetrics=0.5.1=pypi_0
torchvision=0.10.1=py38_cu111
tornado=6.1=pypi_0
tqdm=4.62.3=pypi_0
traitlets=5.1.1=pypi_0
typing_extensions=3.10.0.2=pyh06a4308_0
uncertainty-molecules=0.1=dev_0
urllib3=1.26.7=pypi_0
wandb=0.12.4=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=2.0.2=pypi_0
wheel=0.37.0=pyhd3eb1b0_1
widgetsnbextension=3.5.2=pypi_0
wrapt=1.13.2=pypi_0
x264=1!161.3030=h7f98852_1
xz=5.2.5=h7b6447c_0
yacs=0.1.8=pypi_0
yarl=1.7.0=pypi_0
yaspin=2.1.0=pypi_0
zipp=3.6.0=pypi_0
zlib=1.2.11=h36c2ea0_1013
zstd=1.4.9=haebb681_0
